Move rule compat-uses-vars to own file
